requirements.txt
streamlit>=1.30.0
PyPDF2>=3.0.0
python-docx>=1.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
apscheduler>=3.10.0
pyyaml>=6.0
jinja2>=3.1.0
